// frontend/src/pages/Employment/types/Employee.ts
export type EmployeeDataType = {
	_id: string;
	name: string;
	surname: string;
	fullName: string;
	email: string;
	phoneNumber: number;
	profilePhoto: string;
	salary: number;
	teamLeader: string;
	startingDate: string;
	position: string;
	nID: string;
	status: string;
	gender: string;
	birthDay: Date;
	// status: "Working" | "Remote" | "On Leave";
	contract: string;
};